big changes to backbone view manager, events handling, and routing

+++ bin/start_server.sh
@@ -3,7 +3,7 @@ ulimit -n 9999
 
 # kill existing process
 pid=$(ps aux| grep 'node app.js' |grep -v grep| awk '{print $2}')
-kill -9 $pid
+if [ -n "$pid" ]; then kill -9 $pid; fi
 
 # for dev, use nodemon
 nodemon --watch ./app  --ext '.js|.css|.html|.jinjs'  app.js
